Define slingshot-cxi-drivers-install
slingshot-cxi-drivers-install is a script which can load the CXI drivers
(cxi-ss1, cxi-user, cxi-eth, kfi-cxi). In addition, it supports
reconfiguring the IOMMU group prior to loading the driver.

This script is structure such that it can be run in the modprobe.conf
install cxi-ss1 step.

Udev rules to load the drivers should not be used with this script.

NETCASSINI-7280

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>